+++ octofin/likedplaylist/urls.py
@@ -3,5 +3,5 @@ from .views import jellyfin_webhook, configView
 
 urlpatterns = [
     path('webhook/', jellyfin_webhook, name='jellyfin-webhook'),
-    path('config/', configView, name='likedplaylist-config')
+    path('', configView, name='likedplaylist-config')
 ]